Increase number of startup polls allowed to fail for hang detection
The current value gives Node-RED 37s to start up. We have seen instances when running on slow hardware this isn't quite enough.

This PR doubles the timeout to 75s (10 x 7.5s interval polls). This balances the need to detect a hung node-red in a timely manner vs not letting node-red start at all!