settings: Inject custom catalogues into NR settings
By default this change does only make the catalogue settings as by
default in Node-RED explicit. There's a side-effect when this catalogue
ever changes the default, though ever other instance has to perform an
update too in that case.

When an instance is licensed it has the opportunity to override the
catalogues with a custom set. This would allow the FlowForge application
to control what catalogue is available for the end-user.