Prevent negative CPU usage on restart of NR
Because CPU % is calcualted on the delta in the total process
CPU usage time, on a restart the delta could got very negative.

We now reset the previous value to 0 if the delta is negative and
return a 0% for that sample.

We also reset the last value to zero if no CPU data is found,
This is because that can lead to % greater than 100% as the delta
could be longer than the sample window.